+++ CHANGELOG.md
@@ -4,12 +4,20 @@ All notable changes to this project will be documented in this file. The format
 is roughly based on [Keep a Changelog], and this project tries to adheres to
 [Semantic Versioning].
 
-## [0.1.2-dev] - TBD
+## [0.2.0-dev] - TBD
 
 ### Changed
 
 - Histogam plots in `{ctype}-heuristics.svg` now uses R-style pretty breaks as
   its binning strategy.
+- Moved heuristics thresholds from `config.toml`: `min_f`, `min_tcc`, `min_aq`,
+  `min_conf` are now CLI arguments only with range validation
+- Made `variant_color`, `viz_option`, and `cluster_option` TOML arguments only
+- Made `--paired` option for read-percentages CLI-only. It is required with no
+  default
+- Moved `tree height` to `[cluster_options]` in `config.toml`
+- Enabled all clustermap matrix types by default in `config.toml`: `expenrd`,
+  `jaccard`, `mutuald`, and `njointp`
 
 ## [0.1.1] - 2026-07-16
 
@@ -22,7 +30,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Initial release. IRMA-Viz can reproduce the original IRMA plots.
 
 <!-- Versions -->
-[0.1.2-dev]: https://github.com/CDCgov/irma-viz/compare/v0.1.1...v0.1.2-dev
+[0.2.0-dev]: https://github.com/CDCgov/irma-viz/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/CDCgov/irma-viz/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/CDCgov/irma-viz/releases/tag/v0.1.0